RG-181 Автоматизация. Функция Найти объекты (по устройствам). Некорректная работа условия поиска по АЛС.

+++ Projects/Common/Infrastructure.Automation/ProcedureThread.Steps.cs
@@ -383,7 +383,7 @@ namespace Infrastructure.Automation
 				switch (property)
 				{
 					case Property.ShleifNo:
-						propertyValue = gkDevice.ShleifNo;
+						propertyValue = (int)gkDevice.ShleifNo;
 						break;
 					case Property.IntAddress:
 						propertyValue = gkDevice.IntAddress;
